Move integer register distance code into own file
arm_reg_alloc.c is now pretty much generic.
