+++ Makefile
@@ -22,6 +22,7 @@ ARM =\
 	arm_gen.c \
 	arm_walker.c \
 	arm_reg_alloc.c \
+	arm_int_dist.c \
 	arm_encode.c \
 	arm_link.c \
 	arm2_div.c \
